internal/grpc: split server and resources
